Override user home dir with DROPBEAR_HOME.
As Android is reporting "/" as the home dir for all users, we want to
make sure dropbear looks into DROPBEAR_HOME for files like known_hosts,
public keys etc.
Previously this was overriden by fake/debug getpwnam and getpwuid calls
which are now turned off.
Thus, we now override the home dir globally.
On the other hand, we set the user's $HOME environment variable to "/"
because we want him to browse the tree from its root - as a convenience
measure. He would else end up in DROPBEAR_HOME after logging in and
would have to chdir("/") himself.

Change-Id: Ic84ff4e68d1c1dd4819ca116bbaec96cff4c292d